Rename the biopotential package to eeg
