--- docs/api/interfaces.biopotential.rst
@@ -1,6 +1,0 @@
-Biopotential Signals
-====================
-
-Base MNE Continuous
--------------------
-.. automodule:: neuroconv.datainterfaces.biopotential.basemnecontinuousdatainterface
