+++ docs/api/interfaces.eeg.rst
@@ -1,0 +1,6 @@
+Electroencephalography (EEG, ECoG, sEEG, DBS)
+=============================================
+
+Base MNE Continuous
+-------------------
+.. automodule:: neuroconv.datainterfaces.eeg.basemnecontinuousdatainterface
