+++ src/neuroconv/datainterfaces/eeg/basemnecontinuousdatainterface.py
